@@ -14,12 +14,11 @@ android {
       storePassword keystoreProperties['storePassword']
     }
   }
-  compileSdkVersion 27
-  buildToolsVersion '27.0.3'
+  compileSdkVersion 28
   defaultConfig {
     applicationId "fobo66.exchangecourcesbelarus"
     minSdkVersion 14
-    targetSdkVersion 27
+    targetSdkVersion 28
     versionCode 8
     versionName "1.7"
     multiDexEnabled true
@@ -39,13 +38,12 @@ android {
 }
 
 dependencies {
-  def supportVersion = "27.1.1"
 
-  //design
-  implementation "com.android.support:appcompat-v7:$supportVersion"
-  implementation "com.android.support:cardview-v7:$supportVersion"
-  implementation "com.android.support:recyclerview-v7:$supportVersion"
-  implementation "com.android.support:support-v4:$supportVersion"
+  //androidx
+  implementation 'androidx.appcompat:appcompat:1.0.0'
+  implementation 'androidx.cardview:cardview:1.0.0'
+  implementation 'androidx.recyclerview:recyclerview:1.0.0'
+  implementation 'androidx.legacy:legacy-support-v4:1.0.0'
 
   //location
   implementation "com.google.android.gms:play-services-location:15.0.1"
@@ -53,16 +51,16 @@ dependencies {
   implementation 'org.slf4j:slf4j-nop:1.7.25'
 
   //firebase
-  implementation "com.google.firebase:firebase-core:16.0.1"
-  implementation "com.google.firebase:firebase-database:16.0.1"
-  implementation "com.google.firebase:firebase-perf:16.0.0"
-  implementation 'com.crashlytics.sdk.android:crashlytics:2.9.4'
+  implementation "com.google.firebase:firebase-core:16.0.3"
+  implementation "com.google.firebase:firebase-database:16.0.2"
+  implementation "com.google.firebase:firebase-perf:16.1.0"
+  implementation 'com.crashlytics.sdk.android:crashlytics:2.9.5'
 
   //http
   implementation 'com.squareup.okhttp3:okhttp:3.10.0'
 
   //multidex
-  implementation 'com.android.support:multidex:1.0.3'
+  implementation 'androidx.multidex:multidex:2.0.0'
 
   testImplementation 'junit:junit:4.12'
   testImplementation 'org.mockito:mockito-core:2.8.47'
