@@ -3,11 +3,9 @@ package fobo66.exchangecourcesbelarus.ui;
 import android.app.Dialog;
 import android.content.DialogInterface;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v4.app.DialogFragment;
-
+import androidx.annotation.NonNull;
+import androidx.fragment.app.DialogFragment;
 import com.google.android.gms.common.GoogleApiAvailability;
-
 import fobo66.exchangecourcesbelarus.MainActivity;
 
 /**
