@@ -14,11 +14,10 @@ import android.preference.PreferenceActivity;
 import android.preference.PreferenceFragment;
 import android.preference.PreferenceManager;
 import android.preference.RingtonePreference;
-import android.support.v4.app.NavUtils;
-import android.support.v7.app.ActionBar;
 import android.text.TextUtils;
 import android.view.MenuItem;
-
+import androidx.appcompat.app.ActionBar;
+import androidx.core.app.NavUtils;
 import fobo66.exchangecourcesbelarus.R;
 
 /**
